Move Config to Services and update references
