+++ hugashop/crm/Extensions/BaseExtension.php
@@ -10,7 +10,7 @@
 
 namespace HugaShop\Extensions;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Helper;
 use HugaShop\Models\Settings;
