+++ hugashop/crm/Extensions/FacebookCommerce/Models/FeedGenerator.php
@@ -47,7 +47,7 @@
 namespace HugaShop\Extensions\FacebookCommerce\Models;
 
 use HugaShop\Models\Image;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Helper;
 use HugaShop\Models\Settings;
 use HugaShop\Models\Product\Product;
