+++ hugashop/crm/Extensions/ProductsImport/ProductsImport.php
@@ -11,7 +11,7 @@
 
 namespace HugaShop\Extensions\ProductsImport;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Helper;
 use HugaShop\Services\Request;
