+++ hugashop/crm/Extensions/SeoLinker/SeoLinker.php
@@ -13,7 +13,7 @@
 
 namespace HugaShop\Extensions\SeoLinker;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Request;
 use App\Services\PaginationService;
