+++ hugashop/crm/Extensions/StorageManager/StorageManager.php
@@ -10,7 +10,7 @@
 
 namespace HugaShop\Extensions\StorageManager;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Request;
 use Symfony\Component\Finder\Finder;
