+++ hugashop/crm/Extensions/YandexMerchant/Models/FeedGenerator.php
@@ -16,7 +16,7 @@
 namespace HugaShop\Extensions\YandexMerchant\Models;
 
 use HugaShop\Models\Image;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Helper;
 use HugaShop\Models\Settings;
 use HugaShop\Models\Product\Product;
