+++ hugashop/crm/Models/Order/OrderDelivery.php
@@ -12,7 +12,7 @@
 
 namespace HugaShop\Models\Order;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Helper;
 use HugaShop\Models\BaseModel;
 use HugaShop\Models\Finance\FinancePurse;
