+++ hugashop/crm/Models/Product/ProductBrand.php
@@ -10,7 +10,7 @@
 
 namespace HugaShop\Models\Product;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Helper;
 use HugaShop\Models\BaseModel;
 use Illuminate\Database\Eloquent\Builder;
