+++ hugashop/crm/Models/User/UserNotifier.php
@@ -10,7 +10,7 @@
 
 namespace HugaShop\Models\User;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Helper;
 use HugaShop\Models\BaseModel;
