+++ hugashop/crm/Modules/Delivery/Cdek/Cdek.php
@@ -12,7 +12,7 @@
 namespace HugaShop\Modules\Delivery\Cdek;
 
 use HugaShop\Models\Order\Order;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Models\Order\OrderDelivery;
