+++ hugashop/crm/Modules/Delivery/DeliveryAuto/DeliveryAuto.php
@@ -12,7 +12,7 @@
 namespace HugaShop\Modules\Delivery\DeliveryAuto;
 
 use HugaShop\Models\Order\Order;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Models\Order\OrderDelivery;
