+++ hugashop/crm/Modules/Payment/FopUa/FopUa.php
@@ -19,7 +19,7 @@ namespace HugaShop\Modules\Payment\FopUa;
 use TCPDF;
 use MessageFormatter;
 use IntlDateFormatter;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Helper;
 use HugaShop\Services\Request;
