+++ hugashop/crm/Modules/Payment/IpRu/IpRu.php
@@ -19,7 +19,7 @@ use TCPDF;
 use TCPDF2DBarcode;
 use MessageFormatter;
 use IntlDateFormatter;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Helper;
 use HugaShop\Services\Request;
