+++ hugashop/crm/Modules/Payment/OlxPay/OlxPay.php
@@ -3,7 +3,7 @@
 namespace HugaShop\Modules\Payment\OlxPay;
 
 use HugaShop\Models\Order\Order;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Models\Order\OrderPayment;
 use HugaShop\Models\Finance\FinanceCurrency;
