+++ src/Controller/Admin/Ajax/Export/OrderExport.php
@@ -3,7 +3,7 @@
 namespace App\Controller\Admin\Ajax\Export;
 
 use HugaShop\Models\Order\Order;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Helper;
 use HugaShop\Services\Request;
 use HugaShop\Models\Settings;
