+++ src/Controller/Admin/Ajax/Export/UserExport.php
@@ -3,7 +3,7 @@
 namespace App\Controller\Admin\Ajax\Export;
 
 use HugaShop\Models\User\User;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Request;
 use App\Controller\BaseAdminController;
 use Symfony\Component\Routing\Attribute\Route;
