+++ src/Controller/Admin/Ajax/ImportCsvAjax.php
@@ -12,7 +12,7 @@
 
 namespace App\Controller\Admin\Ajax;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Helper;
 use HugaShop\Models\Product\Product;
 use HugaShop\Services\Request;
