+++ src/Controller/Admin/Ajax/TemplateAjax.php
@@ -2,7 +2,7 @@
 
 namespace App\Controller\Admin\Ajax;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Request;
 use App\Controller\BaseAdminController;
 use Symfony\Component\Routing\Attribute\Route;
