+++ src/Controller/Admin/Order/PaymentController.php
@@ -10,7 +10,7 @@
 
 namespace App\Controller\Admin\Order;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Request;
 use HugaShop\Models\Finance\FinancePurse;
