+++ src/Controller/Admin/Settings/ImagesController.php
@@ -12,7 +12,7 @@
 
 namespace App\Controller\Admin\Settings;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Request;
 use HugaShop\Models\Settings;
