+++ src/Controller/Admin/Settings/SettingsController.php
@@ -12,7 +12,7 @@ namespace App\Controller\Admin\Settings;
 
 use DateTimeZone;
 use HugaShop\Models\Image;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Request;
 use HugaShop\Models\Settings;
