+++ src/Controller/BaseController.php
@@ -10,7 +10,7 @@
 
 namespace App\Controller;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Helper;
 use HugaShop\Services\Design;
 use HugaShop\Services\Request;
