+++ src/Controller/Front/OrderController.php
@@ -10,7 +10,7 @@
 
 namespace App\Controller\Front;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Request;
 use HugaShop\Models\Order\Order;
