+++ src/Controller/Front/SitemapController.php
@@ -10,7 +10,7 @@
 
 namespace App\Controller\Front;
 
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Helper;
 use HugaShop\Models\Product\Product;
 use HugaShop\Models\Content\ContentPost;
