+++ src/Controller/Front/User/UserLoginController.php
@@ -11,7 +11,7 @@
 namespace App\Controller\Front\User;
 
 use HugaShop\Models\User\User;
-use HugaShop\Models\Config;
+use HugaShop\Services\Config;
 use HugaShop\Services\Design;
 use HugaShop\Services\Helper;
 use HugaShop\Services\Request;
